Bug 1353576 Part -1 - Backout changes from beta to central
These are the changes to stub.nsi in order to allow the stub streamlining
patch to apply on central. Streamlining is targeted at beta (53), this will
allow it to be built on central (55) on Oak for testing.
The relevalent changes are primarily backing out 64-bit support, it also
restores some XP and Vista-specific code that was removed.

Also includes backouts to changes in common.nsh, specifically changes to
ElevateUAC, in order to match what will be present in 53.

MozReview-Commit-ID: GeFnmlaD3eZ