Bug 1405396 - Run geckoview tests on all but non-Gradle Android builds. r=gbrown
As we transition to Gradle (and only Gradle) Fennec builds, we're
turning the Gradle builds (Bg) into non-Gradle builds (Bng).  It's not
worth the effort to rename "gradle" to "non-gradle", so this is a
confusing looking patch is turning on GeckoView tests for everything
but Bng (which has kept the "gradle" name).

MozReview-Commit-ID: C1PlsehRwJf

--HG--
extra : rebase_source : 51f15c8ef039db6ee66aea0104a6f5f0e2afab8f