Bug 1411608 - Test IS_NIGHTLY rather than MOZ_UPDATE_CHANNEL. r=Build
The manipulations that determine whether we have a Release, Beta, or
Nightly build are incredibly fragile, and this is one place where
testing the narrow MOZ_UPDATE_CHANNEL for just "nightly" is causing
issues.  On other project branches, MOZ_UPDATE_CHANNEL will be
"nightly-$PROJECT", like "nightly-oak", causing features to not be
compiled.

MozReview-Commit-ID: 29SxsMFb0EX